backend/requirements.txt
Django==5.2.1
gunicorn==23.0.0
dj-database-url==0.5.0
python-decouple==3.8
psycopg2-binary==2.9.9